[string] Use size_t for output length of String::ToCString
String::ToCString returns an off-heap char[] allocated on the C++ heap,
so the buffer's size should use size_t instead of uint32_t. In general
we want to use size_t for the sizes of objects outside of the V8 heap
and uint32_t for sizes of objects inside the V8 heap (for example to
ensure that corruption of size values can not cause an out-of-sandbox
memory access later on).

Bug: 376071292
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6000364
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97101}

+++ src/logging/log.h
@@ -465,10 +465,10 @@ class V8_EXPORT_PRIVATE CodeEventLogger : public LogEventListener {
 
   virtual void LogRecordedBuffer(Tagged<AbstractCode> code,
                                  MaybeHandle<SharedFunctionInfo> maybe_shared,
-                                 const char* name, int length) = 0;
+                                 const char* name, size_t length) = 0;
 #if V8_ENABLE_WEBASSEMBLY
   virtual void LogRecordedBuffer(const wasm::WasmCode* code, const char* name,
-                                 int length) = 0;
+                                 size_t length) = 0;
 #endif  // V8_ENABLE_WEBASSEMBLY
 
   std::unique_ptr<NameBuffer> name_buffer_;
